Add a new locales: Spanish (Paraguay) and Guarani (Paraguay) to locales.php
I am writing to formally request the addition of a new locale for Spanish (Paraguay) and Guarani (Paraguay) on WordPress. This request is to reopen and follow up on the previously initiated issue, which can be found here: Request for Paraguayan Spanish Locale.

The inclusion of the Paraguayan Spanish and guaraní locale is essential to cater to the unique linguistic needs of the Paraguayan WordPress community. This locale will address specific regional variations in vocabulary, syntax, and semantics, providing a more localized and relevant user experience for WordPress users in Paraguay.